refactor(person-medications): split card regions
